Split up `Hyrax::ContainedInWorksBehavior`
In discussion, it was decided that it was better to break this out into
separate modules.
